Finally adding an FC5!
You have waited so long for this... Already prepared for a GUI, this new Factor Calculator edition can be many times faster than the old one, marking one of the biggest-ever performance improvements between commits. This can also sort the list of numbers and automatically checks that numbers are exactly precise, to the single digit! Thanks to these fixes, the FC5 is faster than ever, and I can begin work on versions for the other programming languages. Thanks for this and many other, previous code tricks goes out to the ever-generous people of Stack Exchange.