Temporary upload
I will be changing this drastically because I just realized I didn't think it all the way through. I should be dealing exclusively in values, but have been conflating pointers and the values they point to, which makes reasoning through the programs unnecessarily difficult and causes a lack of intuition in how the data structure member functions should operate.